pyproject.toml
[tool.poetry]
name = "disjoint-mtl"
version = "0.1.0"
description = ""
authors = ["Nik Vaessen <git@mail.vaessen.tech>"]

[tool.poetry.dependencies]
python = "^3.8"
black = { extras = ["d"], version = "^22.8.0" }
psutil = "^5.9.2"
"hurry.filesize" = "^0.9"
torch = "^1.12.1"
data_utility = { path = "../data_utility", develop = true }
pytorch-lightning = "^1.7.6"
wandb = "^0.13.3"
transformers = "^4.22.1"
omegaconf = "^2.2.3"
hydra-core = "^1.2.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
